move the mesh processing functionalities to a separate module
